discord-bot/pyproject.toml
[tool.poetry]
name = "discord-bot"
version = "0.1.0"
description = "Discord bot for Minecraft server control"
authors = ["Melvyn de Kort <melvyn@mdekort.nl>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.3.2"
boto3 = "^1.34.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
moto = "^4.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
